+++ src/scene/common/mesh.cpp
@@ -58,7 +58,7 @@ Mesh loadMesh(std::string defaultTexture, MeshData meshData, std::function<Textu
   // this is directly coupled to :  ./util/loadmodel.h Vertex struct definition
   for (int i = 0; i < NUM_BONES_PER_VERTEX; i++){
     glEnableVertexAttribArray(3 + i);
-    glVertexAttribIPointer(3 + i, 1, GL_BYTE, sizeof(Vertex), (void*) (offsetof(Vertex, boneIndexes) + (sizeof(short) * i)));    
+    glVertexAttribIPointer(3 + i, 1, GL_BYTE, sizeof(Vertex), (void*) (offsetof(Vertex, boneIndexes) + (sizeof(int32_t) * i)));    
 
     glEnableVertexAttribArray(3 + NUM_BONES_PER_VERTEX + i);
     glVertexAttribPointer(3 + NUM_BONES_PER_VERTEX + i, 1, GL_FLOAT, GL_FALSE, sizeof(Vertex), (void*) (offsetof(Vertex, boneWeights) + (sizeof(float) * i)));
